+++ project-egg-core/src/Physics/Vector2D.cpp
@@ -1,10 +1,10 @@
 #include "Vector2D.h"
 
-Vector2D::Vector2D(int x, int y):X(x), Y(y)
+Vector2D::Vector2D(float x, float y):X(x), Y(y)
 {
 }
 
 void Vector2D::LogVector(const char* message)
 {
-	SDL_Log("%s (X, Y) = (%d, %d)\n", message, X, Y);
+	SDL_Log("%s (X, Y) = (%f, %f)\n", message, X, Y);
 }
